refactor: move types to index.ts
- Moved AudioConfig type.
- Moved SpeechOptions type.
- Moved FrameData type.
- Created index.ts for types.
- Improved type organization.
